Skip functional tests when test account4 is unauthorized
Five functional tests rely on swift test account4 being
properly configured in keystone (the tests are already
skipped with tempauth). A regression in devstack appears
to have caused account4 to not be correctly set up which
causes those five tests to fail.

This is a tactical change to skip those tests when the auth
for account4 fails, pending devstack being fixed. Note:

- we do not just skip the tests unconditionally; on a test
  system with correctly setup keystone accounts the tests
  are still run.

- the tests only skip if the auth fails with keystone for
  account4.

Change-Id: Iffef6d54554d1eb71524589cb58a06f874a6afc6
Related-Bug: 1508860