Fix missing txn_id logs in GreenAsyncPile's spawned functions
This commit ensures that the logger thread_locals
value is passed to and set in _get_conn_response methods
executed in a green thread.

Added partial bug tag because in bug description a more
relevant fix is suggested which would fix the bug completely
but for now this makes sense to add this commit for logging.

Co-Authored-By: Clay Gerrard <clay.gerrard@gmail.com>
Change-Id: I13bbf174fdca89318d69bb0674ed23dc9ec25b9a
Partial-Bug: #1409302