Reclaim isolated .meta files
It is possible for an object dir to contain only
a .meta file*, and currently this file will never
be deleted. This patch changes the diskfile on disk
file processing to identify any isolated meta files
and mark them as ready to be reclaimed, so that the
cleanup will reclaim them if they are older than
reclaim age.

*An isolated .meta can occur as follows:

Consider two replicas of the same object whose ondisk files
have diverged due to failures:

  A has t2.ts
  B has t1.data, t4.meta

(The DELETE at t2 did not make it to B. The POST at t4 was
rejected by A.)

After ssync replication the two ondisk file sets will be:

  A has t2.ts
  B has t2.ts, t4.meta

Once t2 becomes older than (current time - reclaim age) we are
left with:

  B has t4.meta

Closes-Bug: #1534685
Change-Id: I3a175fd948a331d08df1b28f9831d7529216c6f1