go: replication prioritizer
This adds two new commands:

  hummingbird moveparts [old ring.gz] [new ring.gz]

After a rebalance, moveparts will diff the two rings and instruct the
object replicators (via HTTP call) to execute jobs that bulk move
any re-assigned partitions.

  hummingbird restoredevice [ip] [device-name]

restoredevice will instruct peers of the selected device to sync over
their copies of any partitions that should be on that drive.  It can
be used to speed up restoring a drive after it's been swapped.

They should both run until all of the jobs have been taken up by the
replicator, so a large job might take a while.

This means the replicator has a semi-real HTTP server now, and not
just a debug port.  It binds to the replication object server's port
+ 500, which is totally arbitrary and open to bikeshedding.

Change-Id: I4bbb2719657eb84c5191e053a1654e986511bb95