go: ObjectEngine abstraction layer
This is a first pass at something analogous to swift's diskfile. The API is
probably not finalized.

There are a few interfaces you have to implement for a new ObjectEngine.
There's also a RegisterObjectEngine function to register a new object engine.
See go/objectserver/objengine.go for information.

I had to rejigger the AtomicFileWriter interface to make them friends, but this
way is fine too.

Change-Id: I032038b5e8a068aecdf8b29631f60ebf00cfb2ee