Wrap and call existing update_footers callbacks
Be nice to other middlewares that might want to set
footers too by calling their callbacks from the encrypter.

Moves the footer metadata construction to the footers callback
so that Etag can be added to the footers after the inner callback
has returned.

Add tests to verify the callback is propagated with and without
the encrypter installing its own callback.

Change-Id: I1b64cfd67a07010c7ba21f31916b73ab567be5a5