crypto - Update container with correct etag on POST
Whatever container update override etag is sent to the
object server with a PUT must be used in container updates
for subsequent POSTs. Unfortunately the update override
headers (x-backend-container-update-override-*) are not
persisted with the object metadata so are not available
when handling a POST.

For EC there is an ugly hack in the object server to
use the x-object-sysmeta-ec-etag when doing a container
update for a POST.

With crypto, the encryption middleware needs to override
the etag (possibly overriding the already overriding EC
etag value) with an encrypted etag value. We therefore
have a similar problem that this override value is not
persisted at the object server.

This patch changes container override headers to use the
x-object-sysmeta-container-update-override- namespace
so that override values are persisted. That then allows
a general mechanism to apply the override values (if any)
from object metadata when constructing a container update
for a POST.

For backwards compatibility the old x-backend-.. style
headers are still supported and the ugly hack for EC etag
override remains in the object server.

Change-Id: Ie258b63a012f0a0cdf9f759e99bbbac1bb6826af