crypto - only set body crypto meta if body is read and encrypted
Currently if the object body has not been read then the
encrypter footers callback will still set crypto_meta
in sysmeta footers. That causes some of the copy-related
functional test failures.

e.g. a PUT with x-copy-from will read its input data
from the source object (in proxy object controller)
and read no data from the PUT request body. That copy
actually works for an encrypted object while we have
a single key - the encrypted object and all its sysmeta
is correctly copied. But the encrypter footers callback
is then called and sets an etag for the empty string
and also sets crypto meta that overwrites the correct
sysmeta crypto_meta that was just copied.

This patch changes the footers callback to only set
crypto related footers if the object body actually
got read and encrypted. Similarly only replace any
Etag in a PUT response if the body was encrypted.

This is not intended to solve the copy issues, it
is simply defensive coding against unknown downstream
behaviour. The Put with x-copy-from example merely
illustrates why it makes sense to not assume the body
is always read.

Change-Id: I0c09cedfabe5093fea367e93652ee89cc44d7d2f