Fix download resumption after getting no data.
When the proxy is handling an object GET response and an object server
fails to send data in a timely fashion, the proxy can pick up where it
left off with another object server; other than a pause in the
download, the client doesn't even know anything happened.

However, if the proxy received the GET response headers but no data,
it would resume at the wrong spot. In particular, for an N-byte
object, it would ask the second object server for the last N-1 bytes
(or equivalently, all but the first byte). For a replicated storage
policy, this would result in the client getting an abbreviated
download, while for an EC storage policy, the proxy would 500 after
trying to decode a fragment set with a bogus fragment in it.

This commit fixes the resumption logic to ask for all N bytes of the
object from the second object server.

Change-Id: Ib9e28c3dceaded1708e7a30844b534566c7a320c