crypto - don't break conditional GETs on unencrypted objects
We handle conditional GETs in the encrypter by (a) encrypting the
If-(None-)Match etag values and adding them to the list of etags
to match and (b) setting X-Backend-Etag-Is-At to point to the
X-Object-Sysmeta-Crypto-Etag value.

Unfortunately if the target object pre-existed encryption being
enabled, or was PUT while crypto was temporarily not in a proxy
pipeline, then there will be no X-Object-Sysmeta-Crypto-Etag in
the object sysmeta.

That's ok for a replication policy since the object server logic
falls back to use the Etag metadata which is the correct value to
match against. But for an EC policy the correct value to match
against in this situation is X-Object-Sysmeta-EC-Etag, which is
not found because the encrypter set X-Backend-Etag-Is-At to point
to X-Object-Sysmeta-Crypto-Etag.

This patch fixes that by making X-Backend-Etag-Is-At be a list of
possible alternate etag locations, to which the encrypter and then
the proxy EC PUT path can append there respective alternate locations.
Then the object server inspects each in turn until a value is found
to match against.

So in this case we would have X-Backend-Etag-Is-At equal to
"X-Object-Sysmeta-Crypto-Etag,X-Object-Sysmeta-Ec-Etag" and for
the unencrypted EC object the object server would find the alternate
etag value at X-Object-Sysmeta-Ec-Etag.

Change-Id: I46ba5b1074a62e57d261e010e1c9df0393ee9609