crypto - put update_content_type back into obj.py
In [1] we pulled update_content_type out of obj.py into
request_helpers.py so that we could use it in the encrypter
to set a content_type before encrypting it.

In [2] we removed content_type encryption, so no longer use
update_content_type in the encrypter, but we did not put the
function back into obj.py.

We should move update_content_type back to obj.py to remove
the unnecessary change to master.

[1] Change-Id: I47f73e4792185cf69afc0b7f82dca5261ad0d6f8
[2] Change-Id: Id0a962f499074c14ff9cae53c1a45a9ea30abc3c

Change-Id: Iaa62b0e189eddf108812ff7b3b7ececdceb7ba82