Put back missing sysmeta unit tests
We lost some unit tests from test_sysmeta.py in [1].
This patch restores them. Since the proxy no longer
implements post-as-copy or COPY then we need to plumb
in a copy middleware to the test app in test_sysmeta.

It is a little odd perhaps to have these tests under
test/unit/proxy but the alternative is to duplicate
the test setup and base test code that already exists
in test_sysmeta.py into test_copy.py.

[1] commit 46d61a4dcd9a5d9157625c06d6fe7d916e80c3d2
    (server side copy middleware)

Change-Id: Iec02387ccbddbe3841a417880389c707cd5c0346