Add Encryption to the default pipeline
This will cause encryption to be "on" by default,
and thus will be exercised in the gate tests.
It is only intended to be included by default while
there is active development on the feature/crypto branch.

Also increases the MAX_HEADER_COUNT in constraints.py
to allow for crypto adding extra headers.

Do not merge this change to the master branch.

Change-Id: I82f874c4d8c2a724c219c804cf456ccc112df3ba