crypto - Don't encrypt zero size object etags
There is no point encrypting the etag of a zero sized object since
the size metadata and the on disk file size reveal the content.
Not doing so avoids the overhead of wrapping a body key, and also
avoids having ciphertext (the encrypted etag) on disk for a known
plaintext (the etag of an empty string).

Change-Id: Idfd0a9a2a7e4a17f8b070f4a1e209c36b503494c