crypto - make some probe tests compatible with crypto
Tests in these modules:

  test_empty_device_handoff.py
  test_object_failures.py
  test_object_handoff.py

previously failed when the crypto middlewares were enabled in the
proxy pipeline. This was because they make assertions about the state
of the on disk object data and etag value based on making direct GETs
to the object servers using direct_client. When crypto is enabled the
on disk content is unknown and not equal to the PUT object plaintext,
so these assertions fail.

We can avoid the failures without needing to require encryption
to be disabled by comparing values of the on disk data "before"
and "after" the test operations rather than comparing "after"
values to an absolute value.

Tests in test_container_merge_policy_index.py passed with crypto
middleware but were using an internal client to put and get objects.
This patch changes those tests to use an external client to put/get
objects so that encryption is used (if installed) and also adds
assertions to check that objects moved by the container reconciler
are not corrupted.

Probe tests in test_container_sync.py may still fail due to the
container sync not having the crypto middlewares configured in
its internal client.

Change-Id: I640c7e738c8a553030d9ed58a953dd3eead0aa07