crypto - stop using key id to derive GET path keys
Currently the key id saved in X-Object-Sysmeta-Crypto-Id is only set
during a PUT request. The sysmeta cannot be set by a POST request. The
key id value is the path used to derive keys and is currently used by
the keymaster to determine what keys should be for decrypter to use in
the GET path. So if we have an object that was PUT without crypto and
then POSTed to with crypto, we have no X-Object-Sysmeta-Crypto-Id
value to derive keys from and cannot therefore decrypt the POSTed
metadata.

We do not actually need to use X-Object-Sysmeta-Crypto-Id to derive
keys - objects do not change location. So in this patch we stop using
it for key derivation in the GET and HEAD paths and instead just use
the path of the GET/HEAD request.

The key id may prove useful in the future if for example an object
were ever relocated, or a keymaster chose a different key with
different id for a POST vs the original PUT. If that were to happen
then we could use the persisted key id to inform the keymaster as to
which key is required for decryption.

For such future-proofing this patch changes the location for
persisting the key id to the crypto_meta dicts for each encrypted
item, and no longer uses the X-Object-Sysmeta-Crypto-Id header. That
way, if ever a different key id were used by a keymaster for a POST vs a
PUT then we'd have the right value stored with the metadata
crypto-meta vs the data crypto-meta.

To enable that, the keymaster now includes the key id in the dict of
results returned from fetch_crypto_keys, so that the encrypter can
embed the key id in crypto meta. The key id is also changed to be a
dict that contains the key derivation path and also a version number
for the keymaster. This is purely for future-proofing - we now persist
sufficient information in crypto-meta to determine which keymaster
version was used and how it derived the keys.

This changes also allows some simplification of the keymaster, and now
the keymaster only calculates keys when they are asked for by a call
to fetch_keys.

In addition, the keymaster will never set the 'swift.crypto.override'
flag anymore. There are still encrypter and decrypter unit tests to
handle the case where somebody uses such a keymaster, but this one
will not make use of that flag.

Change-Id: Icb368305a15e1bbe32483f2e2bbb98a1441a4dad