crypto - remove default root secret
Make it a requirement that root_encryption_secret is
a minimum of 44 base64 encoded characters i.e. a minimum
of 32 bytes encoded to base 64.

This patch still has a default root_secret in the proxy
config.  This is just as temporary as the proxy config
having encryption middleware in the pipeline.  Without this,
functests would fail since devstack has no root secret defined.

Change-Id: I82d183f0b89bfd730578bb64623928bcbfaf657c