Improve linting for third-party middleware
Currently any third-party CI doing linting of their own middlewares
against the Swift codebase don't do so hot -- if they use swob's
Response.status_int in any way, the linter will fail because it only
gets set when status does.

This patch defaults it to None in the class definition, seeing as
a linter can't truly know that status_int will only be created lazily
at runtime, and further, that its existence is tied to setting
Response.status (via __init__ or setting the property directly).

I couldn't actually find a code path within Swift where we could end
up with a None when we access it; at init time, we default the status
code to 200, and there don't seem to be any code paths that init with
a None status code. (You might try, but you can't do that even today.)
As a result, this seems to have no ill effect on anything, but is a
huge help for linters.

This doesn't seem to break any unit tests, func tests via v1, v2, and
v3 auth, nor tempest tests via v2 or v3.

Change-Id: Id28fb56bc7133b9eb59df7989a71b892fd919643