crypto - purge crypto sysmeta from responses
Remove all crypto related (transient-) sysmeta from responses so that
they are not inadvertantly copied to another object. If a COPY is to
an unencrypted destination then the source crypto meta would not apply
and could result in corrupted data when reading the destination.

Also remove container update override headers when copying an object
and the source etag is not trusted (e.g. Range copy or large object
copy).

Drive-by change to stop FakeSwift updating request headers with etag
and content-length that were not in the actual PUT request. That
change was made earlier on this branch to support tests that have now
been deleted.

Change-Id: Ib7cf7676a2f62f79b067f93aa1d2cae64c393fe9