Don't encrypt update override etags for empty object
Fix an anomaly where object metadata for an empty object has no
encrypted etag, but if the encrypter received a container update
override etag in footers or headers then it would encrypt that,
so we'd have encrypted metadata in the container listing but not
in the object metadata. (Empty object etags are not encrypted
because the object content is revealed by its size anyway).

This patch changes the override handling to not encrypt override
etags that correspond to an empty object, with one exception: if
for some reason the received override etag value is that of an empty
string but there *was* an object body, then we'll encrypt the
override etag, because it's value is not obvious from the object
size.

Change-Id: I8d7da34d6d98f351f59174883bc4d5ed0416c101