Add content-encoding parameter to formpost middleware
The content-type parameter can already be set when uploading file
objects, for example as a hidden field in the form sent by the browser.

Sometimes it is also required to set the content-encoding of an object,
for example when uploading a gzipped logfile or html file; see [1] for
further reference. This patch adds this option and tests existing and
new functionality for content-type and content-encoding.

[1] https://tools.ietf.org/html/rfc2616#section-14.11

Change-Id: I3dd47c7c999c29d62075955301fd763ad0b2c602