Add a global fix for anything that involves GetHashPrefixAndSuffix
We tackled this previously with commit
6f78542637b80671cf3153dad8e15ce54dedbb39, but that fix did not
go far enough. Other modules invoke ./hummingbird, so their tests
need the same faking that makeReplicator() needs.

Note that we have 2 kinds of tests: the one that tests the
GetHashPrefixAndSuffix itself and the rest, which uses the
GetHashPrefixAndSuffix in the process of testing something else.
We create a test file for the former and fake everything without
any files for the latter.

Change-Id: Id87cfa6f5271111ae194316de170c6707ba6f23c