Fix inline tempurl/formpost signature examples
The inline examples contain double backslashes, and this creates invalid
signatures. This bugged me a second time now, so let's fix that and make it
fool proof so that I don't need to check the proxy server logs another time for
the same reason.

Changed the docstring to raw docstring to render the backslashes
correctly both in Python and in the restructured text, according to
http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#escaping-mechanism

Change-Id: I77e12cfd12ad998346dd42f7245b9d27124a0a97