Raise ValueError if empty value coming into encrypt_header_val
encrypt_header_val is used to translate a raw header value into an
encrypted value. Semantically, a header with an empty value won't be stored
and all callers seem to remove such a header before calling
encrypted_header_val.

So if no reason for returning ('', None), I think it's better to change it to
raise ValueError not to cause another error for users of the return value.
(e.g. dump_crypto_meta)

Plus this patch addes a few unit tests for those cases above.

Change-Id: Ic1237f4afb8c0e466be5ce59fe31b667c39242b0