Ensure close socket for memcached if got timeout
Currently MemcacheRing attempts to close socket when Timeout occured at
creating connection in MemcacheConnPool. However unfortunately it could
not because if Timeout happened, the sock variable is not assigned and
None value (initial assignment) will be fed into _exception_occured
method.

This patch added a unit test to confirm the socket.close() called and
fixes the behavior with wrapping try/except in MemcacheConnPool.create
to call close() method.

Change-Id: I7959ab96682f4a459a99c4f92463e1280372f6ef