Pass empty etag override values through encrypter
Currently if the encrypter middleware receives an
x-object-sysmeta-container-update-override-etag value of '' in either
a header or footer, it may or not pass the value on to the proxy app:
if there is no object body then the '' value is passed on, but if
there is a body then the override value is replaced with the encrypted
etag of the body.

However strange it may seem for an override etag value to be the empty
string, the encrypter should at least behave consistently, so this
patch makes it always respect the empty string override value and pass
it through, unencrypted (i.e.  with no crypto_meta appended). It is
assumed that the middleware that set the override value knows what it
is doing.

Note that the override etag value is never used to verify the object
content, it is only used to forward a value to the container server
for inclusion with the object's listing entry.

Change-Id: Ie1d073be9ffb00d0f360298c3040eb2f15de0926