Add test for POST to DLO manifest file
In the past, a POST to a DLO manifest file when object_post_as_copy
was true resulted in the manifest file contents being replaced
by the concatenation of the DLO segments.  This no longer
happens, but tests for this case are missing.

This patch adds a functional test to assert that the manifest
file is preserved in a POST request.

Change-Id: I90546014a7dcc7266f0d0e0ff6339688b7954b96
Related-bug: #1487791
Related-bug: #1514317