Deprecate swift-temp-url
python-swiftclient includes an improved and tested method to generate
tempurls. The command syntax is essentially the same, therefore we can
deprecate this one by importing that method.

python-swiftclient is not added as a requirement; if the import fails
due to a missing swiftclient module it will just raise a deprecation
warning.

Closes-Bug: #1607523
Closes-Bug: #1607519

Change-Id: Ifa8bf636f20f82db4845b02d1b58699edaa39356