Use ConfigParser instead of SafeConfigParser
SafeConfigParser supports interpolation on top of ConfigParser in
Python 2, and SafeConfigParser is deprecated in Python 3.2 and log
warning like " DeprecationWarning: The SafeConfigParser class has
been renamed to ConfigParser in Python 3.2. This alias will be removed
in future versions. Use ConfigParser directly instead." So we can use
ConfigParser if we don't need interpolation.

Change-Id: I7e399b3cb90ded909e0d777a4d10c44f1e9299da
Closes-Bug: #1618666